[refactor] #10 - Global Package 정리
