+++ src/main/java/com/dbsgapi/dbsgapi/api/member/controller/MemberApiController.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import com.dbsgapi.dbsgapi.global.security.SecurityUtil;
+import com.dbsgapi.dbsgapi.global.util.SecurityUtil;
 
 import java.util.List;
