+++ src/test/java/com/dbsgapi/dbsgapi/common/KrWonFormatTest.java
@@ -1,6 +1,6 @@
 package com.dbsgapi.dbsgapi.common;
 
-import com.dbsgapi.dbsgapi.global.common.KrWonFormat;
+import com.dbsgapi.dbsgapi.global.util.KrWonFormat;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
